Add guidewire physics module
